Create Task 1
TASK 1 - NUMBER GUESSING GAME
Create a program that generates a random number and asks the user to guess it. Provide feedback on whether the guess is too high or too low until the user guesses the correct number.
Explanation - This code plays a game with the user. It picks a random number between 1 to 10 and the user has five attempts to guess what the correct number is. After each guess , the program tells us whether the guess was TOO HIGH or TOO LOW. If the correct number is guessed it exhibits it so . If the correct number is not guessed even after 5 attempts , it tells you the correct number. 📃